backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-mpv==1.0.7
mutagen==1.47.0
sqlalchemy==2.0.29
aiofiles==23.2.1
pillow==10.3.0
python-multipart==0.0.9
